# packages/api-contracts/openapi/platform-api.yaml
openapi: 3.1.0
info:
  title: Actenora Platform API
  version: 0.2.0
  description: Platform + web-portal contract surface for the modular monolith (FAZ 24).
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/continuity-ledger/suggestions:
    get:
      operationId: listContinuitySuggestions
      summary: List continuity relation suggestions
      responses:
        "200":
          description: Suggestion list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContinuitySuggestion"
    post:
      operationId: recordContinuitySuggestion
      summary: Record a continuity relation suggestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordContinuitySuggestion"
      responses:
        "200":
          description: Recorded suggestion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContinuitySuggestion"
  /api/v1/continuity-ledger/suggestions/{suggestionId}/accept:
    post:
      operationId: acceptContinuitySuggestion
      summary: Accept a continuity relation suggestion
      parameters:
        - name: suggestionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Approved suggestion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContinuitySuggestion"
  /api/v1/continuity-ledger/suggestions/{suggestionId}/reject:
    post:
      operationId: rejectContinuitySuggestion
      summary: Reject a continuity relation suggestion
      parameters:
        - name: suggestionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Rejected suggestion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContinuitySuggestion"
  /api/v1/continuity-ledger/contradictions:
    get:
      operationId: listContinuityContradictions
      summary: List contradiction candidates
      responses:
        "200":
          description: Contradiction list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContinuityContradiction"
    post:
      operationId: proposeContinuityContradiction
      summary: Propose a contradiction candidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposeContinuityContradiction"
      responses:
        "200":
          description: Proposed contradiction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContinuityContradiction"
  /api/v1/continuity-ledger/contradictions/{contradictionId}/confirm:
    post:
      operationId: confirmContinuityContradiction
      summary: Confirm a contradiction candidate
      parameters:
        - name: contradictionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Confirmed contradiction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContinuityContradiction"
  /api/v1/continuity-ledger/contradictions/{contradictionId}/reject:
    post:
      operationId: rejectContinuityContradiction
      summary: Reject a contradiction candidate
      parameters:
        - name: contradictionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Rejected contradiction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContinuityContradiction"
  /api/v1/continuity-ledger/commitments:
    post:
      operationId: recordContinuityCommitment
      summary: Record a commitment in the continuity ledger
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordContinuityCommitment"
      responses:
        "200":
          description: Recorded commitment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContinuityCommitment"
  /api/v1/continuity-ledger/commitments/{commitmentId}/confirm:
    post:
      operationId: confirmContinuityCommitment
      summary: Confirm a pending commitment
      parameters:
        - name: commitmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Confirmed commitment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContinuityCommitment"
  /api/v1/continuity-ledger/commitments/{commitmentId}/reject:
    post:
      operationId: rejectContinuityCommitment
      summary: Reject a pending commitment
      parameters:
        - name: commitmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Rejected commitment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContinuityCommitment"
...
  /api/v1/continuity-ledger/occurrences/{occurrenceId}/brief:
    get:
      operationId: getMeetingBrief
      summary: Generate previous-meeting brief for an occurrence
      parameters:
        - name: occurrenceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Meeting brief
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingBrief"
  /api/v1/continuity-ledger/occurrences/{occurrenceId}/continuity:
    get:
      operationId: getOccurrenceContinuity
      summary: Fetch continuity projection for an occurrence
      parameters:
        - name: occurrenceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Continuity projection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OccurrenceContinuity"
...
  /api/v1/commitments:
    get:
      operationId: listCommitments
      summary: Commitment tracker
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Commitment page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommitmentListResponse"
...
components:
  parameters:
...
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
  schemas:
...
    ArtifactStatus:
      type: string
      enum: [DRAFT, PENDING_APPROVAL, APPROVED, REJECTED, COMPLETED, OPEN, AT RISK]
    EvidenceRef:
      type: object
      additionalProperties: false
      required: [segmentId, startMs, endMs, quote]
      properties:
        segmentId:
          type: string
          format: uuid
        startMs:
          type: integer
        endMs:
          type: integer
        quote:
          type: string
...
    CommitmentItem:
      type: object
      additionalProperties: false
      required: [id, meetingId, statement, ownerDisplayName, dueAt, status, evidence]
      properties:
        id:
          type: string
          format: uuid
        meetingId:
          type: string
          format: uuid
        statement:
          type: string
        ownerDisplayName:
          type: string
        dueAt:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: "#/components/schemas/ArtifactStatus"
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceRef"
...
    ContinuitySuggestion:
      type: object
      additionalProperties: false
      required: [id, sourceOccurrenceId, targetOccurrenceId, proposedRelation, confidence, reason, status]
      properties:
        id:
          type: string
          format: uuid
        sourceOccurrenceId:
          type: string
          format: uuid
        targetOccurrenceId:
          type: string
          format: uuid
        proposedRelation:
          type: string
          enum: [SAME_SERIES, SAME_BUSINESS_CONTEXT, FOLLOW_UP]
        confidence:
          type: number
        reason:
          type: string
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
    RecordContinuitySuggestion:
      type: object
      additionalProperties: false
      required: [sourceOccurrenceId, targetOccurrenceId, proposedRelation, reason]
      properties:
        sourceOccurrenceId:
          type: string
          format: uuid
        targetOccurrenceId:
          type: string
          format: uuid
        proposedRelation:
          type: string
          enum: [SAME_SERIES, SAME_BUSINESS_CONTEXT, FOLLOW_UP]
        confidence:
          type: number
        reason:
          type: string
    ContinuityContradiction:
      type: object
      additionalProperties: false
      required: [id, meetingOccurrenceId, leftDecisionId, rightDecisionId, reason, confidence, status]
      properties:
        id:
          type: string
          format: uuid
        meetingOccurrenceId:
          type: string
          format: uuid
        leftDecisionId:
          type: string
          format: uuid
        rightDecisionId:
          type: string
          format: uuid
        reason:
          type: string
        confidence:
          type: number
        status:
          type: string
          enum: [PENDING, CONFIRMED, REJECTED]
    ProposeContinuityContradiction:
      type: object
      additionalProperties: false
      required: [meetingOccurrenceId, leftDecisionId, rightDecisionId, reason]
      properties:
        meetingOccurrenceId:
          type: string
          format: uuid
        leftDecisionId:
          type: string
          format: uuid
        rightDecisionId:
          type: string
          format: uuid
        reason:
          type: string
        confidence:
          type: number
    ContinuityCommitment:
      type: object
      additionalProperties: false
      required: [commitmentId, meetingOccurrenceId, text, status, overdue]
      properties:
        commitmentId:
          type: string
          format: uuid
        meetingOccurrenceId:
          type: string
          format: uuid
        text:
          type: string
        owner:
          type: string
          nullable: true
        status:
          type: string
        dueDate:
          type: string
          format: date
          nullable: true
        overdue:
          type: boolean
    RecordContinuityCommitment:
      type: object
      additionalProperties: false
      required: [meetingOccurrenceId, text]
      properties:
        meetingOccurrenceId:
          type: string
          format: uuid
        noteId:
          type: string
          format: uuid
        text:
          type: string
        owner:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date
          nullable: true
    MeetingBrief:
      type: object
      additionalProperties: false
      required: [briefId, targetOccurrenceId, openTasks, openRisks, unresolvedQuestions, followUpChain, overdueCommitments, generatedAt]
      properties:
        briefId:
          type: string
          format: uuid
        targetOccurrenceId:
          type: string
          format: uuid
        previousOccurrenceId:
          type: string
          format: uuid
          nullable: true
        meetingSeriesId:
          type: string
          format: uuid
          nullable: true
        openTasks:
          type: array
          items:
            $ref: "#/components/schemas/CarryOverItem"
        openRisks:
          type: array
          items:
            $ref: "#/components/schemas/CarryOverItem"
        unresolvedQuestions:
          type: array
          items:
            $ref: "#/components/schemas/CarryOverItem"
        followUpChain:
          type: array
          items:
            type: string
            format: uuid
        overdueCommitments:
          type: array
          items:
            $ref: "#/components/schemas/ContinuityCommitment"
        generatedAt:
          type: string
          format: date-time
    CarryOverItem:
      type: object
      additionalProperties: false
      required: [itemId, kind, text, sourceOccurrenceId]
      properties:
        itemId:
          type: string
          format: uuid
        kind:
          type: string
        text:
          type: string
        sourceOccurrenceId:
          type: string
          format: uuid
    OccurrenceContinuity:
      type: object
      additionalProperties: false
      required: [meetingOccurrenceId, followUpChain]
      properties:
        meetingOccurrenceId:
          type: string
          format: uuid
        meetingSeriesId:
          type: string
          format: uuid
          nullable: true
        businessContextId:
          type: string
          format: uuid
          nullable: true
        previousOccurrenceId:
          type: string
          format: uuid
          nullable: true
        nextOccurrenceId:
          type: string
          format: uuid
          nullable: true
        followUpChain:
          type: array
          items:
            type: string
            format: uuid
...
    CommitmentListResponse:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CommitmentItem"
        nextCursor:
          type: string
          nullable: true
...